2023 day 11: solve second part by observing much older universe than initially estimated

+++ src/main/java/tk/vivas/adventofcode/year2023/day11/Galaxy.java
@@ -1,8 +1,8 @@
 package tk.vivas.adventofcode.year2023.day11;
 
-record Galaxy(int x, int y) {
+record Galaxy(long x, long y) {
 
-	int distance(Galaxy other) {
+	long distance(Galaxy other) {
 		return Math.abs(other.x - x) + Math.abs(other.y - y);
 	}
 }
